Update downloadCart unit tests - also split downloadApi into api functions and rq hooks - this makes it easier to test and also smaller files - also add test for disabling download of carts with empty items
