+++ packages/datagateway-download/src/downloadApi.test.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { DownloadCartItem, handleICATError } from 'datagateway-common';
+import { handleICATError } from 'datagateway-common';
 import {
   downloadDeleted,
   downloadPreparedCart,
@@ -10,19 +10,7 @@ import {
   fetchAdminDownloads,
   adminDownloadDeleted,
   adminDownloadStatus,
-  useCart,
-  useRemoveAllFromCart,
-  useRemoveEntityFromCart,
-  useIsTwoLevel,
-  useSizes,
-  useDatafileCounts,
 } from './downloadApi';
-import { renderHook, WrapperComponent } from '@testing-library/react-hooks';
-import React from 'react';
-import { createMemoryHistory } from 'history';
-import { QueryClient, QueryClientProvider, setLogger } from 'react-query';
-import { Router } from 'react-router-dom';
-import { DownloadSettingsContext } from './ConfigProvider';
 
 jest.mock('datagateway-common', () => {
   const originalModule = jest.requireActual('datagateway-common');
@@ -56,267 +44,9 @@ const mockedSettings = {
   },
 };
 
-// silence react-query errors
-setLogger({
-  log: console.log,
-  warn: console.warn,
-  error: jest.fn(),
-});
-
-const createTestQueryClient = (): QueryClient =>
-  new QueryClient({
-    defaultOptions: {
-      queries: {
-        retry: false,
-      },
-    },
-  });
-
-const createReactQueryWrapper = (): WrapperComponent<unknown> => {
-  const testQueryClient = createTestQueryClient();
-  const history = createMemoryHistory();
-
-  const wrapper: WrapperComponent<unknown> = ({ children }) => (
-    <DownloadSettingsContext.Provider value={mockedSettings}>
-      <Router history={history}>
-        <QueryClientProvider client={testQueryClient}>
-          {children}
-        </QueryClientProvider>
-      </Router>
-    </DownloadSettingsContext.Provider>
-  );
-  return wrapper;
-};
-
 describe('Download Cart API functions test', () => {
   afterEach(() => {
     (handleICATError as jest.Mock).mockClear();
-  });
-
-  describe('useCart', () => {
-    it('sends axios request to fetch cart and returns successful response', async () => {
-      const downloadCartMockData = {
-        cartItems: [
-          {
-            entityId: 1,
-            entityType: 'investigation',
-            id: 1,
-            name: 'INVESTIGATION 1',
-            parentEntities: [],
-          },
-          {
-            entityId: 2,
-            entityType: 'dataset',
-            id: 2,
-            name: 'DATASET 2',
-            parentEntities: [],
-          },
-        ],
-        createdAt: '2019-11-01T15:18:00Z',
-        facilityName: mockedSettings.facilityName,
-        id: 1,
-        updatedAt: '2019-11-01T15:18:00Z',
-        userName: 'test user',
-      };
-
-      axios.get = jest.fn().mockResolvedValue({
-        data: downloadCartMockData,
-      });
-
-      const { result, waitFor } = renderHook(() => useCart(), {
-        wrapper: createReactQueryWrapper(),
-      });
-
-      await waitFor(() => result.current.isSuccess);
-
-      expect(axios.get).toHaveBeenCalledWith(
-        'https://example.com/downloadApi/user/cart/LILS',
-        {
-          params: {
-            sessionId: null,
-          },
-        }
-      );
-      expect(result.current.data).toEqual(downloadCartMockData.cartItems);
-    });
-
-    it('sends axios request to fetch cart and calls handleICATError on failure', async () => {
-      axios.get = jest.fn().mockRejectedValue({
-        message: 'Test error message',
-      });
-
-      const { result, waitFor } = renderHook(() => useCart(), {
-        wrapper: createReactQueryWrapper(),
-      });
-
-      await waitFor(() => result.current.isError);
-
-      expect(handleICATError).toHaveBeenCalledWith({
-        message: 'Test error message',
-      });
-    });
-  });
-
-  describe('removeAllDownloadCartItems', () => {
-    it('returns nothing upon successful response', async () => {
-      axios.delete = jest.fn().mockImplementation(() =>
-        Promise.resolve({
-          data: {
-            cartItems: [],
-            facilityName: mockedSettings.facilityName,
-            userName: 'test user',
-          },
-        })
-      );
-
-      const { result, waitFor } = renderHook(() => useRemoveAllFromCart(), {
-        wrapper: createReactQueryWrapper(),
-      });
-
-      expect(axios.delete).not.toHaveBeenCalled();
-      expect(result.current.isIdle).toBe(true);
-
-      result.current.mutate();
-
-      await waitFor(() => result.current.isSuccess);
-
-      expect(result.current.data).toBeUndefined();
-      expect(axios.delete).toHaveBeenCalled();
-      expect(
-        axios.delete
-      ).toHaveBeenCalledWith(
-        `${mockedSettings.downloadApiUrl}/user/cart/${mockedSettings.facilityName}/cartItems`,
-        { params: { sessionId: null, items: '*' } }
-      );
-    });
-
-    it('logs error upon unsuccessful response', async () => {
-      axios.delete = jest.fn().mockImplementation(() =>
-        Promise.reject({
-          message: 'Test error message',
-        })
-      );
-
-      const { result, waitFor } = renderHook(() => useRemoveAllFromCart(), {
-        wrapper: createReactQueryWrapper(),
-      });
-
-      result.current.mutate();
-
-      await waitFor(() => result.current.isError);
-
-      expect(
-        axios.delete
-      ).toHaveBeenCalledWith(
-        `${mockedSettings.downloadApiUrl}/user/cart/${mockedSettings.facilityName}/cartItems`,
-        { params: { sessionId: null, items: '*' } }
-      );
-      expect(handleICATError).toHaveBeenCalledWith({
-        message: 'Test error message',
-      });
-    });
-  });
-
-  describe('removeDownloadCartItem', () => {
-    it('returns empty array upon successful response', async () => {
-      axios.delete = jest.fn().mockImplementation(() =>
-        Promise.resolve({
-          data: {
-            cartItems: [],
-            facilityName: mockedSettings.facilityName,
-            userName: 'test user',
-          },
-        })
-      );
-
-      const { result, waitFor } = renderHook(() => useRemoveEntityFromCart(), {
-        wrapper: createReactQueryWrapper(),
-      });
-
-      expect(axios.delete).not.toHaveBeenCalled();
-      expect(result.current.isIdle).toBe(true);
-
-      result.current.mutate({ entityId: 1, entityType: 'datafile' });
-
-      await waitFor(() => result.current.isSuccess);
-
-      expect(result.current.data).toEqual([]);
-      expect(axios.delete).toHaveBeenCalled();
-      expect(
-        axios.delete
-      ).toHaveBeenCalledWith(
-        `${mockedSettings.downloadApiUrl}/user/cart/${mockedSettings.facilityName}/cartItems`,
-        { params: { sessionId: null, items: 'datafile 1' } }
-      );
-    });
-
-    it('logs error upon unsuccessful response', async () => {
-      axios.delete = jest.fn().mockImplementation(() =>
-        Promise.reject({
-          message: 'Test error message',
-        })
-      );
-
-      const { result, waitFor } = renderHook(() => useRemoveEntityFromCart(), {
-        wrapper: createReactQueryWrapper(),
-      });
-
-      result.current.mutate({ entityId: 1, entityType: 'investigation' });
-
-      await waitFor(() => result.current.isError);
-
-      expect(
-        axios.delete
-      ).toHaveBeenCalledWith(
-        `${mockedSettings.downloadApiUrl}/user/cart/${mockedSettings.facilityName}/cartItems`,
-        { params: { sessionId: null, items: 'investigation 1' } }
-      );
-      expect(handleICATError).toHaveBeenCalledWith({
-        message: 'Test error message',
-      });
-    });
-  });
-
-  describe('getIsTwoLevel', () => {
-    it('returns true if IDS is two-level', async () => {
-      axios.get = jest.fn().mockImplementation(() =>
-        Promise.resolve({
-          data: true,
-        })
-      );
-
-      const { result, waitFor } = renderHook(() => useIsTwoLevel(), {
-        wrapper: createReactQueryWrapper(),
-      });
-
-      await waitFor(() => result.current.isSuccess);
-
-      expect(axios.get).toHaveBeenCalledWith(
-        `${mockedSettings.idsUrl}/isTwoLevel`
-      );
-      expect(result.current.data).toEqual(true);
-    });
-
-    it('returns false in the event of an error and logs error upon unsuccessful response', async () => {
-      axios.get = jest.fn().mockImplementation(() =>
-        Promise.reject({
-          message: 'Test error message',
-        })
-      );
-
-      const { result, waitFor } = renderHook(() => useIsTwoLevel(), {
-        wrapper: createReactQueryWrapper(),
-      });
-
-      await waitFor(() => result.current.isError);
-
-      expect(axios.get).toHaveBeenCalledWith(
-        `${mockedSettings.idsUrl}/isTwoLevel`
-      );
-      expect(handleICATError).toHaveBeenCalledWith({
-        message: 'Test error message',
-      });
-    });
   });
 
   describe('submitCart', () => {
@@ -496,218 +226,6 @@ describe('Download Cart API functions test', () => {
       link.target = '_blank';
 
       expect(document.body.appendChild).toHaveBeenCalledWith(link);
-    });
-  });
-
-  describe('useSizes', () => {
-    it('returns the sizes of all the items in a cart', async () => {
-      axios.get = jest
-        .fn()
-        .mockImplementation((path) => {
-          if (path.includes('datafiles/')) {
-            return Promise.resolve({
-              data: {
-                id: 1,
-                name: 'test datafile',
-                fileSize: 1,
-              },
-            });
-          } else {
-            return Promise.resolve({
-              data: 1,
-            });
-          }
-        })
-        .mockImplementationOnce(() =>
-          Promise.reject({
-            message: 'simulating a failed response',
-          })
-        );
-
-      const cartItems: DownloadCartItem[] = [
-        {
-          entityId: 1,
-          entityType: 'investigation',
-          id: 1,
-          name: 'INVESTIGATION 1',
-          parentEntities: [],
-        },
-        {
-          entityId: 2,
-          entityType: 'dataset',
-          id: 2,
-          name: 'DATASET 2',
-          parentEntities: [],
-        },
-        {
-          entityId: 3,
-          entityType: 'datafile',
-          id: 3,
-          name: 'DATAFILE 1',
-          parentEntities: [],
-        },
-        {
-          entityId: 4,
-          entityType: 'investigation',
-          id: 4,
-          name: 'INVESTIGATION 1',
-          parentEntities: [],
-        },
-      ];
-
-      const { result, waitFor } = renderHook(() => useSizes(cartItems), {
-        wrapper: createReactQueryWrapper(),
-      });
-
-      await waitFor(() =>
-        result.current.every((query) => query.isSuccess || query.isError)
-      );
-
-      expect(result.current.map((query) => query.data)).toEqual([
-        undefined,
-        1,
-        1,
-        1,
-      ]);
-      expect(axios.get).toHaveBeenCalledWith(
-        `${mockedSettings.downloadApiUrl}/user/getSize`,
-        {
-          params: {
-            sessionId: null,
-            facilityName: mockedSettings.facilityName,
-            entityType: 'investigation',
-            entityId: 1,
-          },
-        }
-      );
-      expect(axios.get).toHaveBeenCalledWith(
-        `${mockedSettings.downloadApiUrl}/user/getSize`,
-        {
-          params: {
-            sessionId: null,
-            facilityName: mockedSettings.facilityName,
-            entityType: 'dataset',
-            entityId: 2,
-          },
-        }
-      );
-      expect(axios.get).toHaveBeenCalledWith(
-        `${mockedSettings.apiUrl}/datafiles/${3}`,
-        {
-          headers: {
-            Authorization: 'Bearer null',
-          },
-        }
-      );
-      expect(handleICATError).toHaveBeenCalledWith(
-        {
-          message: 'simulating a failed response',
-        },
-        false
-      );
-    });
-  });
-
-  describe('useDatafileCounts', () => {
-    it('returns the counts of all the items in a cart', async () => {
-      axios.get = jest
-        .fn()
-        .mockImplementation(() =>
-          Promise.resolve({
-            data: 1,
-          })
-        )
-        .mockImplementationOnce(() =>
-          Promise.reject({
-            message: 'simulating a failed response',
-          })
-        );
-
-      const cartItems: DownloadCartItem[] = [
-        {
-          entityId: 1,
-          entityType: 'investigation',
-          id: 1,
-          name: 'INVESTIGATION 1',
-          parentEntities: [],
-        },
-        {
-          entityId: 2,
-          entityType: 'investigation',
-          id: 2,
-          name: 'INVESTIGATION 2',
-          parentEntities: [],
-        },
-        {
-          entityId: 3,
-          entityType: 'dataset',
-          id: 3,
-          name: 'DATASET 1',
-          parentEntities: [],
-        },
-        {
-          entityId: 4,
-          entityType: 'datafile',
-          id: 4,
-          name: 'DATAFILE 1',
-          parentEntities: [],
-        },
-      ];
-
-      const { result, waitFor } = renderHook(
-        () => useDatafileCounts(cartItems),
-        {
-          wrapper: createReactQueryWrapper(),
-        }
-      );
-
-      await waitFor(() =>
-        result.current.every((query) => query.isSuccess || query.isError)
-      );
-
-      expect(result.current.map((query) => query.data)).toEqual([
-        undefined,
-        1,
-        1,
-        1,
-      ]);
-      expect(axios.get).toHaveBeenCalledTimes(3);
-      expect(axios.get).toHaveBeenCalledWith(
-        `${mockedSettings.apiUrl}/datafiles/count`,
-        {
-          params: {
-            where: {
-              'dataset.investigation.id': {
-                eq: 2,
-              },
-            },
-          },
-          headers: {
-            Authorization: 'Bearer null',
-          },
-        }
-      );
-      expect(axios.get).toHaveBeenCalledWith(
-        `${mockedSettings.apiUrl}/datafiles/count`,
-        {
-          params: {
-            where: {
-              'dataset.id': {
-                eq: 3,
-              },
-            },
-          },
-          headers: {
-            Authorization: 'Bearer null',
-          },
-        }
-      );
-      expect(handleICATError).toHaveBeenCalledWith(
-        {
-          message: 'simulating a failed response',
-        },
-        false
-      );
     });
   });
 });
